fix: Parsing root_page column of schema
root_page was parsed as u8 which is not correct as acording to docs:
"The maximum page number is 4294967294"
https://www.sqlite.org/fileformat2.html#pages

It could fit in u32 but since the bytes slice length is variable anyway
we parse it to usize for convenience.

The issue did non manifest often as root pages of tables and indexes
were usually at the front, it however caused issues when running
against full companies.db.